requirements.txt
absl-py==2.3.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.2
datasets==3.5.0
decorator==5.2.1
diffusers==0.33.0
dill==0.3.8
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
facexlib==0.3.0
filelock==3.20.1
filterpy==1.4.5
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
grpcio==1.76.0
huggingface-hub==0.30.1
humanfriendly==10.0
ImageIO==2.37.2
imageio-ffmpeg==0.4.2
importlib_metadata==8.7.0
insightface==0.7.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
lpips==0.1.4
Markdown==3.7
MarkupSafe==3.0.3
matplotlib==3.10.3
matplotlib-inline==0.2.1
ml_dtypes==0.5.4
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.13.0
natsort==8.4.0
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
omegaconf==2.0.0
onnx==1.20.0
onnxruntime-gpu==1.22.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.15.2
pexpect==4.9.0
pillow==11.3.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==1.0.8
pytorch-msssim==1.0.0
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
simsimd==6.5.3
stack-data==0.6.3
stringzilla==4.4.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.4
toml==0.10.2
tomli==2.3.0
torch==2.6.0+cu124
torchaudio==2.5.1
torchvision==0.21.0+cu124
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
triton==3.2.0
trl==0.19.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urwid==3.0.4
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.4
xformers==0.0.29.post2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0
